requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
python-dotenv==1.1.1
SQLAlchemy==2.0.23
Werkzeug==3.0.1
gunicorn==23.0.0
groq
numpy